+++ src/components/FilterBarInline/filterBarInlineChangelogData.ts
@@ -18,6 +18,7 @@ export const changelogData: ChangelogData[] = [
       "Removes use of `useNYPLBreakpoints` and replaces with equivalent CSS.",
       "Removes explicit `className` prop as interface can be extended to include Chakra prop or HTML attribute types.",
       "Updates the `id` prop to be optional and generates a random id if not provided.",
+      "Adds `data-testid` value of `ds-filterBarInline` to the parent element.",
     ],
   },
   {