+++ src/components/MultiSelect/MultiSelect.stories.tsx
@@ -319,11 +319,13 @@ export const withControls: Story = {
     buttonText: "MultiSelect",
     closeOnBlur: true,
     id: "multiSelect-with-controls",
+    defaultItemsVisible: undefined,
     isBlockElement: true,
     isDefaultOpen: false,
     isSearchable: true,
     items: withItems,
     listOverflow: "scroll",
+    onClear: undefined,
     onChange: undefined,
     onMixedStateChange: undefined,
     selectedItems: undefined,
@@ -660,6 +662,7 @@ const MultiSelectWithControlsStory = (args) => {
     />
   );
 };
+MultiSelectWithControlsStory.displayName = "MultiSelectWithControlsStory";
 
 const MultiSelectStory = ({
   id,
@@ -720,6 +723,7 @@ const MultiSelectStory = ({
     />
   );
 };
+MultiSelectStory.displayName = "MultiSelectStory";
 
 // TODO: Replace with MultiSelectGroup once that component is done.
 const MultiSelectGroupStory = ({ items }: Partial<MultiSelectProps>) => {
@@ -817,3 +821,4 @@ const MultiSelectGroupStory = ({ items }: Partial<MultiSelectProps>) => {
     </HStack>
   );
 };
+MultiSelectGroupStory.displayName = "MultiSelectGroupStory";