+++ src/components/Tooltip/tooltipChangelogData.ts
@@ -19,6 +19,7 @@ export const changelogData: ChangelogData[] = [
       "Removes explicit `children` prop as its not required when React.PropsWithChildren is used.",
       "Removes `Children.map` logic as warning is no longer going to be used.",
       "Updated the internal use of ids to reduce `undefined` ids in the DOM.",
+      "Adds `data-testid` value of `ds-tooltip` to the parent element.",
     ],
   },
   {